Completed final exercise for Odin CSS Flexbox:
Using 'flex: 1' on a container will force it to use all available space,
which is similar to 'flex-grow: 1' but includes some hidden calcs as
well.

Using 'flex-shrink: 0' will force the container to stay it's size aka it
 will not shrink.

Using 'flex-wrap: wrap' on a container will allow the items within the
container to neatly move to the next line.